api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "RAG Chatbot Backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "fastembed>=0.2.0",
    "qdrant-client>=1.7.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.0.0",
    "tiktoken>=0.5.0",
    "httpx>=0.25.0",
    "google-generativeai>=0.4.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
